Actualizacion y correccion
se rectificaron despues de la migracion, model (entity, request, response), repository (todos los respositorios), services (estandarizacion de nombres)

+++ MsBiblioteca/src/main/java/TDUNU2025/Msbiblioteca/model/entity/Catalogo.java
@@ -14,7 +14,7 @@ public class Catalogo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idCatalogo;
+    private Long idCatalogo;
 
     @Column(nullable = false, length = 100)
     private String nombre;
