+++ MsBiblioteca/src/main/java/TDUNU2025/Msbiblioteca/model/request/AutorRequest.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 
 @Data
 public class AutorRequest {
-    private Integer idPersona;
+    private Long idPersona;
     private String biografia;
     private LocalDate fechaFallecimiento;
     private String fotoUrl;
