+++ MsBiblioteca/src/main/java/TDUNU2025/Msbiblioteca/repository/PrestamoRepository.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Repository;
 import TDUNU2025.Msbiblioteca.model.entity.Prestamo;
 
 @Repository
-public interface PrestamoRepository extends JpaRepository<Prestamo, Integer> {
-    // Heredamos todos los métodos CRUD básicos de JpaRepository:
-    // save(), findById(), findAll(), deleteById(), etc.
+public interface PrestamoRepository extends JpaRepository<Prestamo, Long> {
+    
 }
